fix(budget): qualify compiled contracts and align Helm schemas
Local-only qualification checkpoint on aa34c3fe7ba3f73d26015ab2afe38404c071bb15.
Retain all previously uncommitted parent repairs: existing workspace hex and
schemars dependency edges, kube 3 typed replace_status, ledger borrow/move,
concurrent test root lifetime, required legacy inferenceRef fixture, and real
text/event-stream MIME for all three wiremock SSE fixtures. The original HEAD
was not compile/test qualified; these repairs are part of this checkpoint.

Resolve strict Clippy without new allow/expect attributes, feature flags, or
new dependencies. Rename Amounts.add to checked_add with identical arithmetic.
Keep OutputField operator wire names via explicit serde renames and add exact
serde/JsonSchema compatibility coverage. Move unchanged request normalization,
shape validation and final route/egress classification into a shared router-only
dispatch module, included by the controller only in tests. Keep the public
VerifiedTaskNode.generation field for credential composition and explicitly
verify that generation in the second live inventory, with regression coverage.

Actual Helm/Rust drift tests revealed three pre-existing candidate failures:
Task, Team and Profile budget scope nullability/descriptions, Task's untyped
status budget binding, Team account reference schema, and CEL rule ordering.
Align these templates to the existing generated Rust contract; retain the
same CEL rule strings, lifetime opt-in/retention and legacy launch prohibition.
No fallback prices, raw-token authority, all-compute scope, or feature expansion.

All Cargo commands below used the existing guard, cwd this worktree, and only
/Users/pallakatos/Private/Repos/kars/target. The guard sets CARGO_INCREMENTAL=0,
CARGO_BUILD_JOBS=2, CARGO_NET_OFFLINE=true and the existing CARGO_TARGET_DIR;
it enforces an 8.5 GiB free-space floor. Both packages kept default features.
Exact invocation prefix:
python3 /Users/pallakatos/.copilot/session-state/a7f32227-c055-4df4-81ee-ce27d7f2d74d/files/run-cargo-guard.py --cwd /Users/pallakatos/.copilot/session-state/a7f32227-c055-4df4-81ee-ce27d7f2d74d/kars-pr11-budgets --

Passing guarded commands:
1. cargo test --quiet --offline --locked -p kars-controller -p kars-inference-router --lib --bins -- budget task_identity
   75 controller + 57 router = 132 passed, zero failures/ignored.
   Includes all 16 incomplete Anthropic cases, positive final SSE settlement,
   complete funded-Team interleavings, new generation and wire-schema tests.
2. cargo test --quiet --offline --locked -p kars-controller -p kars-inference-router --lib --bins -- helm_drift::tests::helm_ kars_task::tests:: kars_task::authorization_tests:: kars_team_reconciler::tests:: kars_team_reconciler::persistence_tests:: routes::model_routing routes::chat_completions routes::anthropic_messages routes::inference routes::spawn_policy forward_proxy:: failover:: blocklist::
   68 controller + 83 router = 151 passed, including all 15 Helm drift tests.
   Counts describe executions; some selectors overlap command 1.
3. cargo test --quiet --offline --locked -p kars-controller -p kars-inference-router --test proxy_fake_upstream --test foundry_route_guard --test failover_walk --test multi_provider_guardrails --test anthropic_buffered_guardrail --test chat_output_guardrail_nonjson
   24 real loopback HTTP tests passed: proxy 3, Foundry 9, failover 3,
   multi-provider 5, Anthropic 2, chat non-JSON 2; no external services.
4. cargo clippy --quiet --offline --locked -p kars-controller -p kars-inference-router --all-targets -- -D warnings
   Passed final all-target check, zero warnings and no new waivers.
5. cargo fmt --package kars-controller --package kars-inference-router
6. cargo fmt --package kars-controller --package kars-inference-router --check
   Passed. Only affected files required formatting.
The compatibility/HTTP test invocations set TMPDIR=$PWD/.qualification-fixtures
inside this worktree; that empty scratch directory has been removed.
Minimum observed free space across qualification: 10.03 GiB; floor never hit.

Passing existing Node/static commands, cwd this worktree unless noted:
node cli/node_modules/vitest/vitest.mjs run --root cli --configLoader runner --no-cache src/commands/budget.test.ts
   7 passed. Existing symlinked cache only; no dependency install.
node cli/node_modules/vitest/vitest.mjs run --root deploy/helm/kars/tests --globals --no-cache src/inference-budget.test.ts src/local-inference.test.ts
   18 passed (12 budget + 6 local-inference), rerun after template repair.
cd cli && npm run typecheck
   Passed, no emit.
git diff --check
git diff --cached --check
bash ci/check-copyright-headers.sh
   736 source files passed with the new shared module staged.
BASE_REF=068ae160 bash ci/check-loc.sh
   Passed against the reviewed parent.
Additional source comparison confirmed normalize, shape validation, Quote
accounting, mediated egress and operation classification unchanged modulo
whitespace/relocation; Cargo.lock adds only hex and schemars 1.2.1 package edges.

Limits: this is source/compile/strict-lint/test qualification, not production
acceptance. Real budget/SRE Kind, human review of these repairs, combined
credential/budget authority composition, and an independently fully-qualified
router image digest remain pending. Privacy epoch remains an additional fence,
not Pod UID/custom-audience TokenReview authentication. No public push, merge,
main/customer change, Azure, Docker, image build/push or private Bridge actions.
cli/node_modules remains untracked and is not part of this checkpoint.

Copilot-Session: eb3654cd-f1e0-445a-8734-430800af1903
