+++ deploy/helm/kars/templates/crd-karsprofile.yaml
@@ -59,8 +59,11 @@ spec:
                     properties:
                       scope:
                         type: string
-                        enum: [GovernedInference]
-                        description: "Governed inference only; no all-in task cost claim"
+                        enum: [GovernedInference, null]
+                        nullable: true
+                        description: |-
+                          Explicit opt-in to governed-inference tokens/configured maximum prices.
+                          An absent scope retains the foundation's planning-only interpretation.
                       tokens:
                         description: |-
                           Maximum total tokens the task subtree may consume. `0`/absent means
