+++ deploy/helm/kars/templates/crd-karstask.yaml
@@ -183,8 +183,11 @@ spec:
                     properties:
                       scope:
                         type: string
-                        enum: [GovernedInference]
-                        description: "Explicit governed-inference token/configured maximum-price scope; other costs are excluded"
+                        enum: [GovernedInference, null]
+                        nullable: true
+                        description: |-
+                          Explicit opt-in to governed-inference tokens/configured maximum prices.
+                          An absent scope retains the foundation's planning-only interpretation.
                       tokens:
                         description: |-
                           Maximum total tokens the task subtree may consume. `0`/absent means
@@ -335,14 +338,6 @@ spec:
               reason: FieldValueForbidden
               rule: >-
                 !has(self.execution) || !self.execution.launch || !has(self.envelope.budget) || ((!has(self.envelope.budget.tokens) || self.envelope.budget.tokens == 0) && (!has(self.envelope.budget.usdMicros) || self.envelope.budget.usdMicros == 0)) || (has(self.envelope.budget.scope) && self.envelope.budget.scope == 'GovernedInference')
-            - message: First finite GovernedInference opt-in requires a new Task/Team UID; an existing unbounded runtime cannot be silently converted
-              reason: FieldValueForbidden
-              rule: >-
-                !has(self.envelope.budget) || !has(self.envelope.budget.scope) || self.envelope.budget.scope != 'GovernedInference' || ((!has(self.envelope.budget.tokens) || self.envelope.budget.tokens == 0) && (!has(self.envelope.budget.usdMicros) || self.envelope.budget.usdMicros == 0)) || (has(oldSelf.envelope.budget) && has(oldSelf.envelope.budget.scope) && oldSelf.envelope.budget.scope == 'GovernedInference' && ((has(oldSelf.envelope.budget.tokens) && oldSelf.envelope.budget.tokens > 0) || (has(oldSelf.envelope.budget.usdMicros) && oldSelf.envelope.budget.usdMicros > 0)))
-            - message: GovernedInference scope cannot be removed or changed for an existing UID
-              reason: FieldValueForbidden
-              rule: >-
-                !has(oldSelf.envelope.budget) || !has(oldSelf.envelope.budget.scope) || oldSelf.envelope.budget.scope != 'GovernedInference' || (has(self.envelope.budget) && has(self.envelope.budget.scope) && self.envelope.budget.scope == 'GovernedInference')
             - message: spec.blueprint.isolation must be one of standard, enhanced, confidential
               reason: FieldValueInvalid
               rule: '!has(self.blueprint) || !has(self.blueprint.isolation) || self.blueprint.isolation in [''standard'',''enhanced'',''confidential'']'
@@ -358,14 +353,69 @@ spec:
             - message: spec.blueprint.egress may list at most 32 destinations
               reason: FieldValueInvalid
               rule: '!has(self.blueprint) || !has(self.blueprint.egress) || size(self.blueprint.egress) <= 32'
+            - message: First finite GovernedInference opt-in requires a new Task/Team UID; an existing unbounded runtime cannot be silently converted
+              reason: FieldValueForbidden
+              rule: >-
+                !has(self.envelope.budget) || !has(self.envelope.budget.scope) || self.envelope.budget.scope != 'GovernedInference' || ((!has(self.envelope.budget.tokens) || self.envelope.budget.tokens == 0) && (!has(self.envelope.budget.usdMicros) || self.envelope.budget.usdMicros == 0)) || (has(oldSelf.envelope.budget) && has(oldSelf.envelope.budget.scope) && oldSelf.envelope.budget.scope == 'GovernedInference' && ((has(oldSelf.envelope.budget.tokens) && oldSelf.envelope.budget.tokens > 0) || (has(oldSelf.envelope.budget.usdMicros) && oldSelf.envelope.budget.usdMicros > 0)))
+            - message: GovernedInference scope cannot be removed or changed for an existing UID
+              reason: FieldValueForbidden
+              rule: >-
+                !has(oldSelf.envelope.budget) || !has(oldSelf.envelope.budget.scope) || oldSelf.envelope.budget.scope != 'GovernedInference' || (has(self.envelope.budget) && has(self.envelope.budget.scope) && self.envelope.budget.scope == 'GovernedInference')
           status:
             description: '`KarsTask.status`.'
             nullable: true
             properties:
               inferenceBudget:
                 type: object
-                x-kubernetes-preserve-unknown-fields: true
-                description: "Controller-owned immutable budget account and Task UID ancestry binding"
+                nullable: true
+                description: |-
+                  Controller-owned immutable account/UID ancestry binding. Not a task-name
+                  metering label and never supplied by an agent header.
+                properties:
+                  account:
+                    type: object
+                    properties:
+                      name:
+                        type: string
+                      namespace:
+                        type: string
+                      uid:
+                        type: string
+                    required: [name, namespace, uid]
+                  authorizationDigest:
+                    type: string
+                  parentTaskUid:
+                    type: string
+                    nullable: true
+                  root:
+                    type: object
+                    properties:
+                      clusterUid:
+                        type: string
+                      kind:
+                        type: string
+                        enum: [KarsTask, KarsTeam]
+                      resource:
+                        type: object
+                        properties:
+                          name:
+                            type: string
+                          namespace:
+                            type: string
+                          uid:
+                            type: string
+                        required: [name, namespace, uid]
+                      workspaceUid:
+                        type: string
+                    required: [clusterUid, kind, resource, workspaceUid]
+                  rootTaskUid:
+                    type: string
+                  scope:
+                    type: string
+                    enum: [GovernedInference]
+                  taskUid:
+                    type: string
+                required: [account, authorizationDigest, root, rootTaskUid, scope, taskUid]
               conditions:
                 description: |-
                   Standard K8s conditions. `Ready` is set `True` once the envelope has
