+++ inference-router/Cargo.toml
@@ -25,6 +25,7 @@ futures.workspace = true
 # Serialization
 serde.workspace = true
 serde_json.workspace = true
+schemars.workspace = true
 
 # UTC calendar math for token budget tracker (daily / monthly reset
 # on UTC boundary). Used by `budget.rs` only.
@@ -68,6 +69,7 @@ kars-a2a-core = { path = "../kars-a2a-core", version = "0.1.0" }
 aes-gcm = "0.10"
 hkdf = "0.12"
 sha2 = "0.10"
+hex.workspace = true
 rand = "0.9"
 base64 = "0.22"
 flate2 = "1"
