+++ shared/inference_budget/ledger_tests.rs
@@ -84,7 +84,7 @@ fn request(task: &str, pod: &str, sequence: u64) -> ReserveRequest {
         endpoint: "https://provider.example".into(),
         model: "model".into(),
         operation: Operation::ChatCompletions,
-        output_field: OutputField::MaxTokens,
+        output_field: OutputField::Tokens,
         maximum_input_tokens: 10,
         maximum_output_tokens: 20,
         maximum_wire_bytes: 4096,
